Address review changes
